Fixed Board Indexing, Implemented CountInARow
Fixed incorrect math in CurrBoard representation of the board.

CountInARow can be used to count the number of LIVE 2s, 3s or 4s. To be used to score/rate boards.

To Be Implemented: Minimax searching algorithm.

Also, need to update genMoves to return boards instead of moves so that they can be rated